+++ NEWS.md
@@ -1,3 +1,8 @@
+# rmdfiltr 0.1.4
+
+* Adds character count filter.
+* Adds adapted [`doi2cite`](https://github.com/korintje/pandoc-doi2cite/blob/main/doi2cite.lua) filter and corresponding functions `post_process_doi_citations()` and `replace_resolved_doi_citations`. For details see `vignette("doi2cite", package = "rmdfiltr")`.
+
 # rmdfiltr 0.1.3
 
 * Adds support for `citeproc`.